Update herd_manager tests
- created `tests/animal_module_tests/herd_manager/` folder
- moved the previous herd manager test file into the new folder
- separated the tests in the previous file by functionalities
- wrote new tests for herd manager properties, four herd statistics functions, and fixed tests for daily_routines and _create_newborn_calf
