+++ tests/animal_module_tests/herd_manager/pytest_fixtures.py
@@ -1,0 +1,1043 @@
+from typing import Any
+from unittest import mock
+
+import pytest
+from pytest_mock import MockerFixture
+
+from RUFAS.biophysical.animal.animal import Animal
+from RUFAS.biophysical.animal.data_types.animal_population import AnimalPopulation
+from RUFAS.biophysical.animal.data_types.animal_types import AnimalType
+from RUFAS.biophysical.animal.herd_manager import HerdManager
+from RUFAS.biophysical.animal.nutrients.nutrients import Nutrients
+from RUFAS.biophysical.animal.pen import Pen
+from RUFAS.data_structures.feed_storage_to_animal_connection import Feed
+from RUFAS.enums import AnimalCombination
+from RUFAS.time import Time
+from RUFAS.weather import Weather
+
+
+@pytest.fixture
+def config_json() -> dict[str, Any]:
+    return {
+        "start_date": "2012:244",
+        "end_date": "2018:243",
+        "random_seed": 42,
+        "set_seed": True,
+        "simulate_animals": True,
+        "nutrient_standard": "NASEM",
+        "FIPS_county_code": 55025,
+        "include_detailed_values": True,
+    }
+
+
+@pytest.fixture
+def animal_json() -> dict[str, Any]:
+    return {
+        "herd_information": {
+            "calf_num": 8,
+            "heiferI_num": 44,
+            "heiferII_num": 38,
+            "heiferIII_num_springers": 5,
+            "cow_num": 100,
+            "replace_num": 500,
+            "herd_num": 100,
+            "breed": "HO",
+            "parity_fractions": {"1": 0.386, "2": 0.281, "3": 0.333},
+            "annual_milk_yield": None,
+        },
+        "herd_initialization": {"initial_animal_num": 10000, "simulation_days": 5000},
+        "animal_config": {
+            "management_decisions": {
+                "breeding_start_day_h": 380,
+                "heifer_repro_method": "ED",
+                "cow_repro_method": "ED-TAI",
+                "semen_type": "conventional",
+                "days_in_preg_when_dry": 218,
+                "heifer_repro_cull_time": 500,
+                "do_not_breed_time": 185,
+                "cull_milk_production": 30,
+                "cow_times_milked_per_day": 3,
+                "milk_fat_percent": 4,
+                "milk_protein_percent": 3.2,
+            },
+            "farm_level": {
+                "calf": {
+                    "male_calf_rate_sexed_semen": 0.1,
+                    "male_calf_rate_conventional_semen": 0.53,
+                    "keep_female_calf_rate": 1,
+                    "wean_day": 60,
+                    "wean_length": 7,
+                    "milk_type": "whole",
+                },
+                "repro": {
+                    "voluntary_waiting_period": 50,
+                    "conception_rate_decrease": 0.026,
+                    "decrease_conception_rate_in_rebreeding": False,
+                    "decrease_conception_rate_by_parity": False,
+                    "avg_gestation_len": 276,
+                    "std_gestation_len": 6,
+                    "prefresh_day": 21,
+                    "calving_interval": 400,
+                    "heifers": {
+                        "estrus_detection_rate": 0.9,
+                        "estrus_conception_rate": 0.6,
+                        "repro_sub_protocol": "5dCG2P",
+                        "repro_sub_properties": {"conception_rate": 0.6, "estrus_detection_rate": 0.9},
+                    },
+                    "cows": {
+                        "estrus_detection_rate": 0.6,
+                        "ED_conception_rate": 0.5,
+                        "presynch_program": "Double OvSynch",
+                        "presynch_program_start_day": 50,
+                        "ovsynch_program": "OvSynch 56",
+                        "ovsynch_program_start_day": 64,
+                        "ovsynch_program_conception_rate": 0.6,
+                        "resynch_program": "TAIafterPD",
+                    },
+                },
+                "bodyweight": {
+                    "birth_weight_avg_ho": 43.9,
+                    "birth_weight_std_ho": 1,
+                    "birth_weight_avg_je": 27.2,
+                    "birth_weight_std_je": 1,
+                    "target_heifer_preg_day": 399,
+                    "mature_body_weight_avg": 740.1,
+                    "mature_body_weight_std": 73.5,
+                },
+            },
+            "from_literature": {
+                "repro": {
+                    "preg_check_day_1": 32,
+                    "preg_loss_rate_1": 0.02,
+                    "preg_check_day_2": 60,
+                    "preg_loss_rate_2": 0.096,
+                    "preg_check_day_3": 200,
+                    "preg_loss_rate_3": 0.017,
+                    "avg_estrus_cycle_return": 23,
+                    "std_estrus_cycle_return": 6,
+                    "avg_estrus_cycle_heifer": 21,
+                    "std_estrus_cycle_heifer": 2.5,
+                    "avg_estrus_cycle_cow": 21,
+                    "std_estrus_cycle_cow": 4,
+                    "avg_estrus_cycle_after_pgf": 5,
+                    "std_estrus_cycle_after_pgf": 2,
+                },
+                "culling": {
+                    "cull_day_count": [0, 5, 15, 45, 90, 135, 180, 225, 270, 330, 380, 430, 480, 530],
+                    "feet_leg_cull": {
+                        "probability": 0.1633,
+                        "cull_day_prob": [0, 0.03, 0.08, 0.16, 0.25, 0.36, 0.48, 0.59, 0.69, 0.78, 0.85, 0.90, 0.95, 1],
+                    },
+                    "injury_cull": {
+                        "probability": 0.2883,
+                        "cull_day_prob": [0, 0.08, 0.18, 0.28, 0.38, 0.47, 0.56, 0.64, 0.71, 0.78, 0.85, 0.90, 0.95, 1],
+                    },
+                    "mastitis_cull": {
+                        "probability": 0.2439,
+                        "cull_day_prob": [0, 0.06, 0.12, 0.19, 0.30, 0.43, 0.56, 0.68, 0.78, 0.85, 0.90, 0.94, 0.97, 1],
+                    },
+                    "disease_cull": {
+                        "probability": 0.1391,
+                        "cull_day_prob": [0, 0.04, 0.12, 0.24, 0.34, 0.42, 0.50, 0.57, 0.64, 0.72, 0.81, 0.89, 0.95, 1],
+                    },
+                    "udder_cull": {
+                        "probability": 0.0645,
+                        "cull_day_prob": [0, 0.12, 0.24, 0.33, 0.41, 0.48, 0.55, 0.62, 0.68, 0.76, 0.82, 0.89, 0.95, 1],
+                    },
+                    "unknown_cull": {
+                        "probability": 0.1009,
+                        "cull_day_prob": [0, 0.05, 0.11, 0.18, 0.27, 0.37, 0.45, 0.54, 0.62, 0.70, 0.77, 0.84, 0.92, 1],
+                    },
+                    "parity_death_prob": [0.039, 0.056, 0.085, 0.117],
+                    "parity_cull_prob": [0.169, 0.233, 0.301, 0.408],
+                    "death_day_prob": [0, 0.18, 0.32, 0.42, 0.48, 0.54, 0.60, 0.65, 0.70, 0.77, 0.83, 0.89, 0.95, 1],
+                },
+                "life_cycle": {"still_birth_rate": 0.065},
+            },
+        },
+        "methane_mitigation": {
+            "methane_mitigation_method": "None",
+            "methane_mitigation_additive_amount": 0,
+            "3-NOP_additive_amount": 70,
+            "monensin_additive_amount": 24,
+            "essential_oils_additive_amount": 0,
+            "seaweed_additive_amount": 0,
+        },
+        "housing": "barn",
+        "pasture_concentrate": 0,
+        "methane_model": "IPCC",
+        "ration": {"phosphorus_requirement_buffer": 0, "user_input": False, "formulation_interval": 30},
+        "pen_information": [
+            {
+                "id": 0,
+                "pen_name": "",
+                "animal_combination": "CALF",
+                "vertical_dist_to_milking_parlor": 0.1,
+                "horizontal_dist_to_milking_parlor": 10,
+                "number_of_stalls": 30,
+                "housing_type": "open air barn",
+                "pen_type": "freestall",
+                "max_stocking_density": 1.2,
+                "manure_management_scenario_id": 0,
+            },
+            {
+                "id": 1,
+                "pen_name": "",
+                "animal_combination": "GROWING",
+                "vertical_dist_to_milking_parlor": 0.1,
+                "horizontal_dist_to_milking_parlor": 10,
+                "number_of_stalls": 125,
+                "housing_type": "open air barn",
+                "pen_type": "freestall",
+                "max_stocking_density": 1.2,
+                "manure_management_scenario_id": 1,
+            },
+            {
+                "id": 2,
+                "pen_name": "",
+                "animal_combination": "CLOSE_UP",
+                "vertical_dist_to_milking_parlor": 0.1,
+                "horizontal_dist_to_milking_parlor": 10,
+                "number_of_stalls": 60,
+                "housing_type": "open air barn",
+                "pen_type": "freestall",
+                "max_stocking_density": 1.2,
+                "manure_management_scenario_id": 2,
+            },
+            {
+                "id": 3,
+                "pen_name": "",
+                "animal_combination": "LAC_COW",
+                "vertical_dist_to_milking_parlor": 0.1,
+                "horizontal_dist_to_milking_parlor": 10,
+                "number_of_stalls": 150,
+                "housing_type": "open air barn",
+                "pen_type": "freestall",
+                "max_stocking_density": 1.2,
+                "manure_management_scenario_id": 5,
+            },
+        ],
+    }
+
+
+@pytest.fixture
+def manure_management_json() -> dict[str, Any]:
+    return {
+        "manure_management_scenarios": [
+            {
+                "scenario_id": 0,
+                "bedding_type": "sawdust",
+                "manure_handler": "manual scraping",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "anaerobic lagoon",
+            },
+            {
+                "scenario_id": 1,
+                "bedding_type": "sawdust",
+                "manure_handler": "manual scraping",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "slurry storage outdoor",
+            },
+            {
+                "scenario_id": 2,
+                "bedding_type": "sawdust",
+                "manure_handler": "manual scraping",
+                "manure_separator": "screw press",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "slurry storage outdoor",
+            },
+            {
+                "scenario_id": 3,
+                "bedding_type": "sawdust",
+                "manure_handler": "flush system",
+                "manure_separator": "rotary screen",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "anaerobic lagoon",
+            },
+            {
+                "scenario_id": 4,
+                "bedding_type": "sand",
+                "manure_handler": "flush system",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "anaerobic lagoon",
+            },
+            {
+                "scenario_id": 5,
+                "bedding_type": "sawdust",
+                "manure_handler": "manual scraping",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "anaerobic digestion and lagoon",
+            },
+            {
+                "scenario_id": 6,
+                "bedding_type": "sawdust",
+                "manure_handler": "flush system",
+                "manure_separator": "rotary screen",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "anaerobic digestion and lagoon",
+            },
+            {
+                "scenario_id": 7,
+                "bedding_type": "sawdust",
+                "manure_handler": "flush system",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "rotary screen",
+                "manure_treatment": "anaerobic digestion and lagoon with separator",
+            },
+            {
+                "scenario_id": 8,
+                "bedding_type": "CBPB sawdust",
+                "manure_handler": "tillage",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "compost bedded pack barn",
+            },
+            {
+                "scenario_id": 9,
+                "bedding_type": "none",
+                "manure_handler": "harrowing",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "open lots",
+            },
+            {
+                "scenario_id": 10,
+                "bedding_type": "none",
+                "manure_handler": "flush system",
+                "manure_separator": "screw press",
+                "manure_separator_after_digestion": "rotary screen",
+                "manure_treatment": "anaerobic digestion and lagoon with separator",
+            },
+            {
+                "scenario_id": 11,
+                "bedding_type": "sawdust",
+                "manure_handler": "alley scraper",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "composting",
+            },
+            {
+                "scenario_id": 12,
+                "bedding_type": "straw",
+                "manure_handler": "manual scraping",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "composting",
+            },
+            {
+                "scenario_id": 13,
+                "bedding_type": "straw",
+                "manure_handler": "alley scraper",
+                "manure_separator": "none",
+                "manure_separator_after_digestion": "none",
+                "manure_treatment": "composting",
+            },
+        ],
+        "bedding_configs": [
+            {
+                "name": "sawdust",
+                "bedding_type": "sawdust",
+                "bedding_mass_per_day": 1.97,
+                "bedding_density": 250.0,
+                "bedding_dry_matter_content": 0.9,
+                "bedding_cleaned_fraction": 1.0,
+                "bedding_carbon_fraction": 0.0,
+                "bedding_phosphorus_content": 0.0,
+                "sand_removal_efficiency": 0.0,
+            },
+            {
+                "name": "CBPB sawdust",
+                "bedding_type": "CBPB sawdust",
+                "bedding_mass_per_day": 12,
+                "bedding_density": 350.0,
+                "bedding_dry_matter_content": 0.9,
+                "bedding_cleaned_fraction": 1.0,
+                "bedding_carbon_fraction": 0.35,
+                "bedding_phosphorus_content": 0.0,
+                "sand_removal_efficiency": 0.0,
+            },
+            {
+                "name": "manure solids",
+                "bedding_type": "manure solids",
+                "bedding_mass_per_day": 2.50,
+                "bedding_density": 400.0,
+                "bedding_dry_matter_content": 0.9,
+                "bedding_cleaned_fraction": 1.0,
+                "bedding_carbon_fraction": 0.0,
+                "bedding_phosphorus_content": 0.0,
+                "sand_removal_efficiency": 0.0,
+            },
+            {
+                "name": "straw",
+                "bedding_type": "straw",
+                "bedding_mass_per_day": 1.97,
+                "bedding_density": 100.0,
+                "bedding_dry_matter_content": 0.9,
+                "bedding_cleaned_fraction": 1.0,
+                "bedding_carbon_fraction": 0.35,
+                "bedding_phosphorus_content": 0.0,
+                "sand_removal_efficiency": 0.0,
+            },
+            {
+                "name": "sand",
+                "bedding_type": "sand",
+                "bedding_mass_per_day": 25.0,
+                "bedding_density": 1500.0,
+                "bedding_dry_matter_content": 0.9,
+                "bedding_cleaned_fraction": 1.0,
+                "bedding_carbon_fraction": 0.0,
+                "bedding_phosphorus_content": 0.0,
+                "sand_removal_efficiency": 1.0,
+            },
+            {
+                "name": "none",
+                "bedding_type": "none",
+                "bedding_mass_per_day": 0.0,
+                "bedding_density": 0.0,
+                "bedding_dry_matter_content": 0.0,
+                "bedding_cleaned_fraction": 0.0,
+                "bedding_carbon_fraction": 0.0,
+                "bedding_phosphorus_content": 0.0,
+                "sand_removal_efficiency": 0.0,
+            },
+        ],
+        "manure_handler_configs": [
+            {
+                "name": "flush system",
+                "manure_handler_type": "flush system",
+                "cleaning_water_use_rate": 757.0,
+                "minutes_per_cleaning": 8,
+                "cleanings_per_day": 2,
+                "daily_tillage_frequency": 0,
+                "cleaning_water_recycle_fraction": 0.80,
+            },
+            {
+                "name": "manual scraping",
+                "manure_handler_type": "manual scraping",
+                "cleaning_water_use_rate": 10.0,
+                "minutes_per_cleaning": 8,
+                "cleanings_per_day": 2,
+                "daily_tillage_frequency": 0,
+                "cleaning_water_recycle_fraction": 0.10,
+            },
+            {
+                "name": "alley scraper",
+                "manure_handler_type": "alley scraper",
+                "cleaning_water_use_rate": 10.0,
+                "minutes_per_cleaning": 8,
+                "cleanings_per_day": 2,
+                "daily_tillage_frequency": 0,
+                "cleaning_water_recycle_fraction": 0.10,
+            },
+            {
+                "name": "tillage",
+                "manure_handler_type": "tillage",
+                "cleaning_water_use_rate": 0.0,
+                "minutes_per_cleaning": 0.0,
+                "cleanings_per_day": 0,
+                "daily_tillage_frequency": 1,
+                "cleaning_water_recycle_fraction": 0.0,
+            },
+            {
+                "name": "harrowing",
+                "manure_handler_type": "harrowing",
+                "cleaning_water_use_rate": 0.0,
+                "minutes_per_cleaning": 0.0,
+                "cleanings_per_day": 0,
+                "daily_tillage_frequency": 0,
+                "cleaning_water_recycle_fraction": 0.0,
+            },
+        ],
+        "manure_separator_configs": [
+            {
+                "name": "rotary screen",
+                "manure_separator_type": "rotary screen",
+                "percent_dry_solids": 0.20,
+                "total_solids_removal_efficiency_for_separator": 0.35,
+                "volatile_solids_removal_efficiency_for_separator": 0.40,
+                "nitrogen_removal_efficiency_for_separator": 0.30,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_separator": 0.15,
+                "phosphorus_removal_efficiency_for_separator": 0.40,
+                "potassium_removal_efficiency_for_separator": 0.15,
+            },
+            {
+                "name": "screw press",
+                "manure_separator_type": "screw press",
+                "percent_dry_solids": 0.35,
+                "total_solids_removal_efficiency_for_separator": 0.25,
+                "volatile_solids_removal_efficiency_for_separator": 0.30,
+                "nitrogen_removal_efficiency_for_separator": 0.30,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_separator": 0.10,
+                "phosphorus_removal_efficiency_for_separator": 0.20,
+                "potassium_removal_efficiency_for_separator": 0.23,
+            },
+        ],
+        "manure_treatment_configs": [
+            {
+                "name": "slurry storage underfloor",
+                "manure_treatment_type": "slurry storage underfloor",
+                "total_solids_removal_efficiency_for_treatment": 0.0,
+                "volatile_solids_removal_efficiency_for_treatment": 0.0,
+                "nitrogen_removal_efficiency_for_treatment": 0.0,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_treatment": 0.0,
+                "phosphorus_removal_efficiency_for_treatment": 0.0,
+                "potassium_removal_efficiency_for_treatment": 0.0,
+                "storage_time_period": 120,
+                "manure_cover": "crust",
+            },
+            {
+                "name": "slurry storage outdoor",
+                "manure_treatment_type": "slurry storage outdoor",
+                "total_solids_removal_efficiency_for_treatment": 0.00,
+                "volatile_solids_removal_efficiency_for_treatment": 0.0,
+                "nitrogen_removal_efficiency_for_treatment": 0.0,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_treatment": 0.0,
+                "phosphorus_removal_efficiency_for_treatment": 0.00,
+                "potassium_removal_efficiency_for_treatment": 0.00,
+                "storage_time_period": 120,
+                "freeboard_input": 0.3048,
+                "manure_cover": "crust",
+            },
+            {
+                "name": "anaerobic lagoon",
+                "manure_treatment_type": "anaerobic lagoon",
+                "total_solids_removal_efficiency_for_treatment": 0.0,
+                "volatile_solids_removal_efficiency_for_treatment": 0.0,
+                "nitrogen_removal_efficiency_for_treatment": 0.0,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_treatment": 0.0,
+                "phosphorus_removal_efficiency_for_treatment": 0.0,
+                "potassium_removal_efficiency_for_treatment": 0.0,
+                "hydraulic_retention_time": 365,
+                "sludge_accumulation_period": 10.0,
+                "sludge_accumulation_volume_fraction": 0.00,
+                "storage_time_period": 365,
+                "freeboard_input": 0.3048,
+                "manure_cover": "cover",
+            },
+            {
+                "name": "anaerobic digestion",
+                "manure_treatment_type": "anaerobic digestion",
+                "total_solids_removal_efficiency_for_treatment": 0.45,
+                "volatile_solids_removal_efficiency_for_treatment": 0.40,
+                "nitrogen_removal_efficiency_for_treatment": 0.0,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_treatment": 0.1,
+                "phosphorus_removal_efficiency_for_treatment": 0.0,
+                "potassium_removal_efficiency_for_treatment": 0.0,
+                "hydraulic_retention_time": 25,
+                "sludge_accumulation_period": 1.0,
+                "sludge_accumulation_volume_fraction": 0.03,
+                "top_cover_volume_fraction": 0.2,
+                "evaporation_fraction": 0.02,
+                "anaerobic_digestion_temperature_set_point": 37.5,
+                "anaerobic_digestion_temperature_celsius": 37.5,
+                "manure_cover": "N/A",
+                "digester_methane_leakage_fraction": 0.01,
+            },
+            {
+                "name": "compost bedded pack barn",
+                "manure_treatment_type": "compost bedded pack barn",
+                "total_solids_removal_efficiency_for_treatment": 0.0,
+                "nitrogen_removal_efficiency_for_treatment": 0.0,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_treatment": 0.0,
+                "phosphorus_removal_efficiency_for_treatment": 0.0,
+                "potassium_removal_efficiency_for_treatment": 0.0,
+                "volatile_solids_removal_efficiency_for_treatment": 0.0,
+                "hydraulic_retention_time": 365,
+                "sludge_accumulation_period": 10.0,
+                "sludge_accumulation_volume_fraction": 0.00251,
+                "storage_time_period": 120,
+                "freeboard_input": 0.3048,
+            },
+            {
+                "name": "open lots",
+                "manure_treatment_type": "open lots",
+                "total_solids_removal_efficiency_for_treatment": 0.0,
+                "nitrogen_removal_efficiency_for_treatment": 0.0,
+                "total_ammoniacal_nitrogen_removal_efficiency_for_treatment": 0.0,
+                "phosphorus_removal_efficiency_for_treatment": 0.0,
+                "potassium_removal_efficiency_for_treatment": 0.0,
+                "volatile_solids_removal_efficiency_for_treatment": 0.0,
+                "storage_time_period": 365,
+            },
+            {
+                "name": "composting",
+                "manure_treatment_type": "composting",
+                "composting_type": "intensive windrow",
+                "storage_time_period": 365,
+            },
+        ],
+    }
+
+@pytest.fixture
+def feed_json() -> dict[str, Any]:
+    return {
+    "calf_feeds": [
+        202,
+        216
+    ],
+    "growing_feeds": [
+        2,
+        44,
+        51,
+        110,
+        167,
+        176,
+        231,
+        234
+    ],
+    "close_up_feeds": [
+        2,
+        44,
+        51,
+        100,
+        110,
+        167,
+        231,
+        234
+    ],
+    "lac_cow_feeds": [
+        2,
+        44,
+        51,
+        94,
+        110,
+        167,
+        231,
+        234
+    ],
+    "purchased_feeds": [
+        {
+            "purchased_feed": 2,
+            "purchased_feed_cost": 0.154
+        },
+        {
+            "purchased_feed": 44,
+            "purchased_feed_cost": 0.208
+        },
+        {
+            "purchased_feed": 51,
+            "purchased_feed_cost": 0.005
+        },
+        {
+            "purchased_feed": 94,
+            "purchased_feed_cost": 0.01
+        },
+        {
+            "purchased_feed": 100,
+            "purchased_feed_cost": 0.005
+        },
+        {
+            "purchased_feed": 110,
+            "purchased_feed_cost": 0.005
+        },
+        {
+            "purchased_feed": 167,
+            "purchased_feed_cost": 0.489
+        },
+        {
+            "purchased_feed": 176,
+            "purchased_feed_cost": 0.005
+        },
+        {
+            "purchased_feed": 202,
+            "purchased_feed_cost": 0.001
+        },
+        {
+            "purchased_feed": 216,
+            "purchased_feed_cost": 1.0
+        },
+        {
+            "purchased_feed": 231,
+            "purchased_feed_cost": 0.794
+        },
+        {
+            "purchased_feed": 234,
+            "purchased_feed_cost": 0.331
+        }
+    ],
+    "farm_grown_feeds": [
+        1
+    ],
+    "storage_options": [
+        {
+            "storage_type": "Bunker Silo",
+            "moisture": "Direct Cut",
+            "additive": "preservative",
+            "packing_density": 200,
+            "inoculation": "heterofermentative",
+            "bunk_type": "open_floor",
+            "ventilation": True,
+            "removal_rate": 6,
+            "initial_dry_matter": 0
+        },
+        {
+            "storage_type": "Bunker Silo",
+            "moisture": "Direct Cut",
+            "additive": "preservative",
+            "packing_density": 200,
+            "inoculation": "heterofermentative",
+            "bunk_type": "open_floor",
+            "ventilation": True,
+            "removal_rate": 6,
+            "initial_dry_matter": 0
+        }
+    ],
+    "user_defined_ration_percentages": {
+        "calf": [
+            {
+                "feed_type": 202,
+                "ration_percentage": 50
+            },
+            {
+                "feed_type": 216,
+                "ration_percentage": 50
+            }
+        ],
+        "growing": [
+            {
+                "feed_type": 2,
+                "ration_percentage": 3.4
+            },
+            {
+                "feed_type": 44,
+                "ration_percentage": 4.2
+            },
+            {
+                "feed_type": 51,
+                "ration_percentage": 30.8
+            },
+            {
+                "feed_type": 110,
+                "ration_percentage": 36.3
+            },
+            {
+                "feed_type": 167,
+                "ration_percentage": 4.8
+            },
+            {
+                "feed_type": 176,
+                "ration_percentage": 17.5
+            },
+            {
+                "feed_type": 231,
+                "ration_percentage": 1.5
+            },
+            {
+                "feed_type": 234,
+                "ration_percentage": 1.5
+            }
+        ],
+        "close_up": [
+            {
+                "feed_type": 2,
+                "ration_percentage": 3.8
+            },
+            {
+                "feed_type": 44,
+                "ration_percentage": 4.1
+            },
+            {
+                "feed_type": 51,
+                "ration_percentage": 40.6
+            },
+            {
+                "feed_type": 94,
+                "ration_percentage": 22.2
+            },
+            {
+                "feed_type": 110,
+                "ration_percentage": 21.5
+            },
+            {
+                "feed_type": 167,
+                "ration_percentage": 5.7
+            },
+            {
+                "feed_type": 231,
+                "ration_percentage": 1.1
+            },
+            {
+                "feed_type": 234,
+                "ration_percentage": 1.0
+            }
+        ],
+        "lac_cow": [
+            {
+                "feed_type": 2,
+                "ration_percentage": 11.7
+            },
+            {
+                "feed_type": 44,
+                "ration_percentage": 11.2
+            },
+            {
+                "feed_type": 51,
+                "ration_percentage": 39.3
+            },
+            {
+                "feed_type": 100,
+                "ration_percentage": 7.5
+            },
+            {
+                "feed_type": 110,
+                "ration_percentage": 13.7
+            },
+            {
+                "feed_type": 167,
+                "ration_percentage": 13.1
+            },
+            {
+                "feed_type": 231,
+                "ration_percentage": 1.8
+            },
+            {
+                "feed_type": 234,
+                "ration_percentage": 1.7
+            }
+        ],
+        "tolerance": 0.1,
+        "milk_reduction_maximum": 0.5
+    },
+        "allowances": [
+            {
+                "purchased_feed": 2,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 44,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 51,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 94,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 100,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 110,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 167,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 176,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 202,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 216,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 231,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            },
+            {
+                "purchased_feed": 234,
+                "runtime_purchase_allowance": 1000.0,
+                "advance_purchase_allowance": 1000.0,
+                "planning_cycle_allowance": 1000.0
+            }
+        ]
+}
+
+@pytest.fixture
+def mock_get_data_side_effect(
+    config_json: dict[str, Any],
+        animal_json: dict[str, Any],
+        manure_management_json: dict[str, Any],
+        feed_json: dict[str, Any]
+) -> list[Any]:
+    return [config_json, animal_json, manure_management_json, feed_json, feed_json["allowances"]]
+
+
+@pytest.fixture
+def mock_herd(mocker: MockerFixture) -> dict[str, list[Animal]]:
+    calves = [
+        mock_animal(AnimalType.CALF, mocker, id=0),
+        mock_animal(AnimalType.CALF, mocker, id=1),
+        mock_animal(AnimalType.CALF, mocker, id=2),
+    ]
+    heiferIs = [
+        mock_animal(AnimalType.HEIFER_I, mocker, id=3),
+        mock_animal(AnimalType.HEIFER_I, mocker, id=4),
+        mock_animal(AnimalType.HEIFER_I, mocker, id=5),
+    ]
+    heiferIIs = [
+        mock_animal(AnimalType.HEIFER_II, mocker, id=6),
+        mock_animal(AnimalType.HEIFER_II, mocker, id=7),
+        mock_animal(AnimalType.HEIFER_II, mocker, id=8),
+    ]
+    heiferIIIs = [
+        mock_animal(AnimalType.HEIFER_III, mocker, id=9),
+        mock_animal(AnimalType.HEIFER_III, mocker, id=10),
+        mock_animal(AnimalType.HEIFER_III, mocker, id=11),
+        mock_animal(AnimalType.HEIFER_III, mocker, id=12),
+    ]
+    dry_cows = [
+        mock_animal(AnimalType.DRY_COW, mocker, days_in_milk=0, days_in_pregnancy=0, id=13),
+        mock_animal(AnimalType.DRY_COW, mocker, days_in_milk=0, days_in_pregnancy=10, id=14),
+        mock_animal(AnimalType.DRY_COW, mocker, days_in_milk=0, days_in_pregnancy=50, id=15),
+    ]
+    lac_cows = [
+        mock_animal(AnimalType.LAC_COW, mocker, id=16),
+        mock_animal(AnimalType.LAC_COW, mocker, id=17),
+        mock_animal(AnimalType.LAC_COW, mocker, id=18),
+    ]
+    replacement = [mock_animal(AnimalType.HEIFER_III, mocker, id=19)]
+
+    return {
+        "calves": calves,
+        "heiferIs": heiferIs,
+        "heiferIIs": heiferIIs,
+        "heiferIIIs": heiferIIIs,
+        "dry_cows": dry_cows,
+        "lac_cows": lac_cows,
+        "replacement": replacement,
+    }
+
+
+def mock_herd_manager(
+    calves: list[Animal],
+    heiferIs: list[Animal],
+    heiferIIs: list[Animal],
+    heiferIIIs: list[Animal],
+    cows: list[Animal],
+    replacement: list[Animal],
+    mocker: MockerFixture,
+    mock_get_data_side_effect: list[Any],
+) -> tuple[HerdManager, dict[str, mock.MagicMock]]:
+    mock_feed = mocker.MagicMock(auto_spec=Feed)
+    mock_weather = mocker.MagicMock(auto_spec=Weather)
+    mock_time = mocker.MagicMock(auto_spec=Time)
+    mock_available_feeds = [mock_feed] * 8
+
+    mock_get_data = mocker.patch("RUFAS.input_manager.InputManager.get_data", side_effect=mock_get_data_side_effect)
+    mock_initialize_animal_config = mocker.patch(
+        "RUFAS.biophysical.animal.animal_config.AnimalConfig.initialize_animal_config"
+    )
+    mock_set_lactation_parameters = mocker.patch(
+        "RUFAS.biophysical.animal.milk.lactation_curve.LactationCurve.set_lactation_parameters"
+    )
+    mock_set_milk_quality = mocker.patch(
+        "RUFAS.biophysical.animal.milk.milk_production.MilkProduction.set_milk_quality"
+    )
+    # mock_user_defined_ration_init = mocker.patch(
+    #     "RUFAS.biophysical.animal.ration.user_defined_ration_manager.UserDefinedRationManager.set_user_defined_rations",
+    #     return_value=None
+    # )
+    mocker.patch("RUFAS.data_structures.feed_storage_to_animal_connection.AdvancePurchaseAllowance.__init__",
+                 return_value=None)
+    mocker.patch("RUFAS.biophysical.animal.pen.Pen.update_animals", return_value=None)
+    mock_herd_factory_init = mocker.patch(
+        "RUFAS.biophysical.animal.herd_factory.HerdFactory.__init__", return_value=None
+    )
+    mock_initialize_herd = mocker.patch(
+        "RUFAS.biophysical.animal.herd_factory.HerdFactory.initialize_herd",
+        return_value=AnimalPopulation(
+            calves=calves,
+            heiferIs=heiferIs,
+            heiferIIs=heiferIIs,
+            heiferIIIs=heiferIIIs,
+            cows=cows,
+            replacement=replacement),
+    )
+    mock_purchased_feed_emissions_estimator_init = mocker.patch(
+        "RUFAS.routines.animal.purchased_feed_emissions_estimator.PurchasedFeedEmissionsEstimator.__init__",
+        return_value=None,
+    )
+
+    herd_manager = HerdManager(mock_weather, mock_time, True, mock_available_feeds)
+    # herd_manager.all_pens = []
+
+    return herd_manager, {
+        "mock_get_data": mock_get_data,
+        "mock_initialize_animal_config": mock_initialize_animal_config,
+        "mock_set_lactation_parameters": mock_set_lactation_parameters,\
+        "mock_herd_factory_init": mock_herd_factory_init,
+        "mock_initialize_herd": mock_initialize_herd,
+        "mock_purchased_feed_emissions_estimator_init": mock_purchased_feed_emissions_estimator_init,
+    }
+
+
+def mock_animal(
+        animal_type: AnimalType,
+        mocker: MockerFixture,
+        days_in_milk: int = 0,
+        days_in_pregnancy: int = 0,
+        id: int = 0,
+        body_weight: float = 88.8,
+        total_phosphorus: float = 18.8,
+        sold: bool = False
+) -> Animal:
+    animal = mocker.MagicMock(auto_spec=Animal)
+    animal.id = id
+    animal.animal_type = animal_type
+    if animal_type.is_cow:
+        animal.is_milking = True if animal_type == AnimalType.LAC_COW else False
+        animal.days_in_milk = days_in_milk
+    animal.is_pregnant = True if days_in_milk > 0 else False
+    animal.days_in_pregnancy = days_in_pregnancy
+    animal.body_weight = body_weight
+    animal.nutrients = mocker.MagicMock(auto_spec=Nutrients)
+    animal.nutrients.total_phosphorus_in_animal = total_phosphorus
+    animal.sold = sold
+    return animal
+
+
+def mock_pen(animal_combination: AnimalCombination, mocker: MockerFixture) -> Pen:
+    pen = mocker.MagicMock(auto_spec=Pen)
+    pen.animal_combination = animal_combination
+    return pen
+
+
+@pytest.fixture
+def herd_manager(mock_herd: dict[str, list[Animal]], mock_get_data_side_effect: list[Any], mocker: MockerFixture) -> HerdManager:
+    herd_manager, _ = mock_herd_manager(
+        calves=mock_herd["calves"],
+        heiferIs=mock_herd["heiferIs"],
+        heiferIIs=mock_herd["heiferIIs"],
+        heiferIIIs=mock_herd["heiferIIIs"],
+        cows=mock_herd["dry_cows"] + mock_herd["lac_cows"],
+        replacement=mock_herd["replacement"],
+        mocker=mocker,
+        mock_get_data_side_effect=mock_get_data_side_effect,
+    )
+    return herd_manager
